added @Lazy for cycled injection worked CreditTask

+++ backend/src/main/java/pl/sda/poznan/bank/backend/model/BankAccount.java
@@ -23,12 +23,12 @@ public class BankAccount {
 
     private double balance;
 
-    private  int accountNumber;
+    private  long accountNumber;
 
     public BankAccount() {
     }
 
-    public BankAccount(AccountType accountType, User user, int accountNumber) {
+    public BankAccount(AccountType accountType, User user, long accountNumber) {
         this.accountType = accountType;
         this.user = user;
         this.accountNumber = accountNumber;
